requirements.txt
crewai
openai
python-dotenv
pydantic
streamlit
langchain
langchain-openai
requests
beautifulsoup4
selenium
pandas
crewai_tools
langchain-google-genai